requirements.txt
backcall==0.1.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
Django==2.0
django-angular==2.1
django-filter==2.0.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
oauthlib==2.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.0
PyJWT==1.6.4
pytz==2018.7
requests==2.21.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
